8265891: ChannelInputStream.transferTo() uses FileChannel.transferTo()
As suggested in 8265891 this change implements a performance optimization
for ChannelInputStream.transferTo(): It uses FileChannel.transferTo()
in case the wrapped channel actually is a FileChannel, as in that case,
the actual work of transferring bytes can potentially get offloaded to the
operating system / file system.

As more such optimizations are likely to follow in the very same code
segment, this change is deliberately elaborate and rather complex in its
design, e. g. it already holds blocking locks for the complete time of
operation on input and output channels.

Signed-off-by: Markus Karg <markus@headcrashing.eu>